+++ crypto/evp_extra/p_kem.c
@@ -8,7 +8,7 @@
 
 #include "../fipsmodule/evp/internal.h"
 #include "../fipsmodule/delocate.h"
-#include "../kem/internal.h"
+#include "../fipsmodule/kem/internal.h"
 #include "../internal.h"
 #include "internal.h"
