+++ crypto/fipsmodule/bcm.c
@@ -120,6 +120,7 @@
 #include "hmac/hmac.c"
 #include "kdf/kbkdf.c"
 #include "kdf/sskdf.c"
+#include "kem/kem.c"
 #include "md4/md4.c"
 #include "md5/md5.c"
 #include "ml_kem/ml_kem.c"
