+++ crypto/kem/kem_methods.c
@@ -123,117 +123,3 @@ const KEM_METHOD kem_kyber1024r3_method = {
   kyber1024r3_encaps,
   kyber1024r3_decaps,
 };
-
-static int ml_kem_512_keygen_deterministic(uint8_t *public_key,
-                                               uint8_t *secret_key,
-                                               const uint8_t *seed) {
-  return ml_kem_512_keypair_deterministic(public_key, secret_key, seed) == 0;
-}
-
-static int ml_kem_512_keygen(uint8_t *public_key,
-                             uint8_t *secret_key) {
-  return ml_kem_512_keypair(public_key, secret_key) == 0;
-}
-
-static int ml_kem_512_encaps_deterministic(uint8_t *ciphertext,
-                                               uint8_t *shared_secret,
-                                               const uint8_t *public_key,
-                                               const uint8_t *seed) {
-  return ml_kem_512_encapsulate_deterministic(ciphertext, shared_secret, public_key, seed) == 0;
-}
-
-static int ml_kem_512_encaps(uint8_t *ciphertext,
-                             uint8_t *shared_secret,
-                             const uint8_t *public_key) {
-  return ml_kem_512_encapsulate(ciphertext, shared_secret, public_key) == 0;
-}
-
-static int ml_kem_512_decaps(uint8_t *shared_secret,
-                             const uint8_t *ciphertext,
-                             const uint8_t *secret_key) {
-  return ml_kem_512_decapsulate(shared_secret, ciphertext, secret_key) == 0;
-}
-
-const KEM_METHOD kem_ml_kem_512_method = {
-  ml_kem_512_keygen_deterministic,
-  ml_kem_512_keygen,
-  ml_kem_512_encaps_deterministic,
-  ml_kem_512_encaps,
-  ml_kem_512_decaps,
-};
-
-static int ml_kem_768_keygen_deterministic(uint8_t *public_key,
-                                               uint8_t *secret_key,
-                                               const uint8_t *seed) {
-  return ml_kem_768_keypair_deterministic(public_key, secret_key, seed) == 0;
-}
-
-static int ml_kem_768_keygen(uint8_t *public_key,
-                                 uint8_t *secret_key) {
-  return ml_kem_768_keypair(public_key, secret_key) == 0;
-}
-
-static int ml_kem_768_encaps_deterministic(uint8_t *ciphertext,
-                                               uint8_t *shared_secret,
-                                               const uint8_t *public_key,
-                                               const uint8_t *seed) {
-  return ml_kem_768_encapsulate_deterministic(ciphertext, shared_secret, public_key, seed) == 0;
-}
-
-static int ml_kem_768_encaps(uint8_t *ciphertext,
-                                 uint8_t *shared_secret,
-                                 const uint8_t *public_key) {
-  return ml_kem_768_encapsulate(ciphertext, shared_secret, public_key) == 0;
-}
-
-static int ml_kem_768_decaps(uint8_t *shared_secret,
-                                 const uint8_t *ciphertext,
-                                 const uint8_t *secret_key) {
-  return ml_kem_768_decapsulate(shared_secret, ciphertext, secret_key) == 0;
-}
-
-const KEM_METHOD kem_ml_kem_768_method = {
-    ml_kem_768_keygen_deterministic,
-    ml_kem_768_keygen,
-    ml_kem_768_encaps_deterministic,
-    ml_kem_768_encaps,
-    ml_kem_768_decaps,
-};
-
-static int ml_kem_1024_keygen_deterministic(uint8_t *public_key,
-                                                uint8_t *secret_key,
-                                                const uint8_t *seed) {
-  return ml_kem_1024_keypair_deterministic(public_key, secret_key, seed) == 0;
-}
-
-static int ml_kem_1024_keygen(uint8_t *public_key,
-                                  uint8_t *secret_key) {
-  return ml_kem_1024_keypair(public_key, secret_key) == 0;
-}
-
-static int ml_kem_1024_encaps_deterministic(uint8_t *ciphertext,
-                                                uint8_t *shared_secret,
-                                                const uint8_t *public_key,
-                                                const uint8_t *seed) {
-  return ml_kem_1024_encapsulate_deterministic(ciphertext, shared_secret, public_key, seed) == 0;
-}
-
-static int ml_kem_1024_encaps(uint8_t *ciphertext,
-                                  uint8_t *shared_secret,
-                                  const uint8_t *public_key) {
-  return ml_kem_1024_encapsulate(ciphertext, shared_secret, public_key) == 0;
-}
-
-static int ml_kem_1024_decaps(uint8_t *shared_secret,
-                                  const uint8_t *ciphertext,
-                                  const uint8_t *secret_key) {
-  return ml_kem_1024_decapsulate(shared_secret, ciphertext, secret_key) == 0;
-}
-
-const KEM_METHOD kem_ml_kem_1024_method = {
-    ml_kem_1024_keygen_deterministic,
-    ml_kem_1024_keygen,
-    ml_kem_1024_encaps_deterministic,
-    ml_kem_1024_encaps,
-    ml_kem_1024_decaps,
-};
